requirements.txt
setuptools>=68.2.2
wheel>=0.38.4
tts==0.22.0
playsound>=1.3.0
loguru>=0.7.2
python-dotenv==1.0.0
transformers==4.36.2
twitchio==2.8.2